docs: generalize documentation for multi-provider support, remove dead code
- Remove unused getTodayDate, MONTH_OFFSET, PAD_LENGTH
- Extract shared SearchHit/StructuredSearchResponse to types.ts
- Make README/AGENTS.md provider-agnostic with supported providers table
- Fix max-statements (20) and max-params (5) values in AGENTS.md
- Add openai keyword to package.json
- Disable import/prefer-default-export lint rule
