+++ .oxlintrc.json
@@ -19,6 +19,7 @@
     "unicorn/prefer-ternary": "off",
     "import/no-nodejs-modules": "off",
     "import/no-named-export": "off",
+    "import/prefer-default-export": "off",
     "import/consistent-type-specifier-style": "off",
     "typescript/consistent-type-imports": "off",
     "max-statements": ["warn", { "max": 20 }],
