+++ src/constants.ts
@@ -4,9 +4,7 @@ const ANTHROPIC_NPM_PACKAGE = "@ai-sdk/anthropic";
 const EMPTY_LENGTH = 0;
 const MAX_RESPONSE_TOKENS = 16_000;
 const MIN_QUERY_LENGTH = 2;
-const MONTH_OFFSET = 1;
 const OPENAI_NPM_PACKAGE = "@ai-sdk/openai";
-const PAD_LENGTH = 2;
 const SEARCH_SYSTEM_PROMPT = "You are an assistant for performing a web search tool use";
 const WEBSEARCH_ALWAYS = "always";
 const WEBSEARCH_AUTO = "auto";
@@ -16,9 +14,7 @@ export {
   EMPTY_LENGTH,
   MAX_RESPONSE_TOKENS,
   MIN_QUERY_LENGTH,
-  MONTH_OFFSET,
   OPENAI_NPM_PACKAGE,
-  PAD_LENGTH,
   SEARCH_SYSTEM_PROMPT,
   WEBSEARCH_ALWAYS,
   WEBSEARCH_AUTO,
